refactor(ref_finance): standardize pool_id as u32
Change pool_id type from u64 to u32 in TokenPairLike trait,
TokenPair implementation, and SwapAction struct for consistency.

+++ backend/src/ref_finance/pool_info.rs
@@ -22,7 +22,7 @@ const POOL_KIND_SIMPLE: &str = "SIMPLE_POOL";
 /// TokenPair の機能を抽象化するトレイト
 pub trait TokenPairLike {
     /// プールIDを返す
-    fn pool_id(&self) -> u64;
+    fn pool_id(&self) -> u32;
 
     /// 入力トークンIDを返す
     fn token_in_id(&self) -> TokenInAccount;
@@ -115,8 +115,8 @@ pub struct TokenPair {
 
 // TokenPairLike トレイトの実装
 impl TokenPairLike for TokenPair {
-    fn pool_id(&self) -> u64 {
-        self.pool.id as u64
+    fn pool_id(&self) -> u32 {
+        self.pool.id 
     }
 
     fn token_in_id(&self) -> TokenInAccount {
